implemented new dashboard
